fix: reuse http client for externaldata requests
