+++ constraint/pkg/externaldata/request_test.go
@@ -1,11 +1,8 @@
 package externaldata
 
 import (
-	"crypto/tls"
 	"reflect"
 	"testing"
-
-	"github.com/open-policy-agent/frameworks/constraint/pkg/apis/externaldata/unversioned"
 )
 
 func TestNewProviderRequest(t *testing.T) {
@@ -61,78 +58,6 @@ func TestNewProviderRequest(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := NewProviderRequest(tt.args.keys); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("NewProviderRequest() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getClient(t *testing.T) {
-	type args struct {
-		provider   *unversioned.Provider
-		clientCert *tls.Certificate
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "invalid http url",
-			args: args{
-				provider: &unversioned.Provider{
-					Spec: unversioned.ProviderSpec{
-						URL: "http://foo",
-					},
-				},
-				clientCert: nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "no CA bundle",
-			args: args{
-				provider: &unversioned.Provider{
-					Spec: unversioned.ProviderSpec{
-						URL: "https://foo",
-					},
-				},
-				clientCert: nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "invalid CA bundle",
-			args: args{
-				provider: &unversioned.Provider{
-					Spec: unversioned.ProviderSpec{
-						URL:      "https://foo",
-						CABundle: badCABundle,
-					},
-				},
-				clientCert: nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "valid CA bundle",
-			args: args{
-				provider: &unversioned.Provider{
-					Spec: unversioned.ProviderSpec{
-						URL:      "https://foo",
-						CABundle: validCABundle,
-					},
-				},
-				clientCert: nil,
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := getClient(tt.args.provider, tt.args.clientCert)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getClient() error = %v, wantErr %v", err, tt.wantErr)
-				return
 			}
 		})
 	}
